archive_and_fixes/requirements_file.txt
# Requirements for IL vs RL Surgical Action Prediction
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
tqdm>=4.62.0

# Transformers and NLP
transformers>=4.20.0

# RL libraries (optional - install if doing RL experiments)
gymnasium>=0.26.0
stable-baselines3>=2.0.0

# Data processing
opencv-python>=4.5.0
pillow>=8.0.0
scipy>=1.7.0

# Logging and utilities
tensorboard>=2.8.0
wandb>=0.12.0  # optional
pyyaml>=6.0
pathlib2>=2.3.0

# Development tools
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0

# Additional utilities
jupyterlab>=3.0.0  # for notebooks
ipywidgets>=7.6.0
